Move around files.
