+++ src/AFQMC/Drivers/AFQMCDriver.cpp
@@ -7,7 +7,7 @@
 #include "OhmmsData/ParameterSet.h"
 #include "OhmmsData/libxmldefs.h"
 #include "Configuration.h"
-#include <qmc_common.h>
+#include <Utilities/qmc_common.h>
 
 #include "AFQMC/config.h"
 #include "AFQMC/Drivers/AFQMCDriver.h"
