+++ src/AFQMC/Drivers/BenchmarkDriver.cpp
@@ -11,7 +11,7 @@
 #include "Message/CommOperators.h"
 #include "OhmmsData/libxmldefs.h"
 #include "Configuration.h"
-#include <qmc_common.h>
+#include <Utilities/qmc_common.h>
 
 #include "AFQMC/config.h"
 #include "AFQMC/Drivers/BenchmarkDriver.h"
