+++ src/Particle/LongRange/KContainer.cpp
@@ -14,7 +14,7 @@
 
 #include "Message/Communicate.h"
 #include "LongRange/KContainer.h"
-#include <qmc_common.h>
+#include <Utilities/qmc_common.h>
 #include <map>
 
 namespace qmcplusplus
