+++ src/QMCApp/CoupledMC.cpp
@@ -26,7 +26,7 @@
 //#include "QMCDrivers/CMC/DummyIonMove.h"
 #include "Message/Communicate.h"
 #include "Message/OpenMP.h"
-#include "qmc_common.h"
+#include "Utilities/qmc_common.h"
 #include "OhmmsData/AttributeSet.h"
 
 namespace qmcplusplus
