+++ src/QMCDrivers/QMCDriverFactory.cpp
@@ -24,7 +24,7 @@
 
 
 #include "QMCDrivers/MCPopulation.h"
-#include "qmc_common.h"
+#include "Utilities/qmc_common.h"
 #include "QMCDrivers/QMCDriverFactory.h"
 #include "QMCWaveFunctions/WaveFunctionPool.h"
 #include "QMCHamiltonians/HamiltonianPool.h"
