+++ src/QMCDrivers/QMCDriverNew.cpp
@@ -26,8 +26,8 @@
 #include "Message/CommOperators.h"
 #include "OhmmsApp/RandomNumberControl.h"
 #include "Estimators/EstimatorManagerNew.h"
-#include "HDFVersion.h"
-#include "qmc_common.h"
+#include "io/HDFVersion.h"
+#include "Utilities/qmc_common.h"
 #include "Concurrency/Info.hpp"
 #include "QMCDrivers/GreenFunctionModifiers/DriftModifierBuilder.h"
 #include "Utilities/StlPrettyPrint.hpp"
