+++ src/QMCHamiltonians/ECPComponentBuilder.cpp
@@ -22,7 +22,7 @@
 #include "Utilities/SimpleParser.h"
 #include "Message/CommOperators.h"
 #include <cmath>
-#include <qmc_common.h>
+#include <Utilities/qmc_common.h>
 
 
 namespace qmcplusplus
