+++ src/QMCWaveFunctions/Jastrow/kSpaceJastrowBuilder.cpp
@@ -15,7 +15,7 @@
 
 #include "QMCWaveFunctions/Jastrow/kSpaceJastrowBuilder.h"
 #include "OhmmsData/AttributeSet.h"
-#include <qmc_common.h>
+#include <Utilities/qmc_common.h>
 
 namespace qmcplusplus
 {
