+++ src/Utilities/qmc_common.h
